+++ server/controller/missing/Missings.ts
@@ -1,4 +1,4 @@
-import prisma from "../client";
+import prisma from "../../client";
 import { Request, Response } from "express";
 import { StatusCodes } from "http-status-codes";
